% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwrappers-package.R
\docType{package}
\name{bwrappers-package}
\alias{bwrappers}
\alias{bwrappers-package}
\title{bwrappers: Clean, analyze, and visualize data in the behavioral sciences}
\description{
This package contains wrappers and other functions for more
    efficiently cleaning, analyzing, and visualizing data in the behavioral
    sciences. The functions do so by:
 \enumerate{
  \item Tailoring syntax to the structure of a typical data sheet in the behavioral sciences
  \item Automating intermediate steps such as splitting and reshaping data frames
  \item Returning statistics in APA format and copying these outputs to the clipboard
 }
}
\references{
\url{https://github.com/michaelkardas/bwrappers}
}
\author{
\strong{Maintainer}: Michael Kardas \email{mkardas@chicagobooth.edu}

}
\keyword{internal}
