% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.anova.R
\name{wrap.anova}
\alias{wrap.anova}
\title{ANOVA (main effects & interaction effects)}
\usage{
wrap.anova(dv1, iv1 = NULL, iv2 = NULL, iv3 = NULL, type = 3)
}
\arguments{
\item{dv1}{Column vector containing the between-subjects dependent variable
OR multiple column vectors containing the within-subjects dependent variables}

\item{iv1, iv2, iv3}{Column vectors containing the between-subjects independent variables}

\item{type}{Numeric argument representing sum-of-squares type (\code{1}, \code{2}, or \code{3})}
}
\description{
Computes main effects and interaction effects for ANOVA with
up to 1 within-subjects factor and up to 3 between-subjects factors. The function
delegates the primary computations to \code{\link[ez]{ezANOVA}}. Note that this
function assumes categorical (i.e., unordered) independent variables, fixed effects,
equality of variances for between-subjects factors, and sphericity for within-subjects
factors. If variance differs significantly by condition in a fully between-subjects
analysis, or if sphericity does not hold, the function additionally displays
assumption checks. In the output, hp2 denotes partial eta squared.
}
\examples{
## ANOVA with 1 within-subjects factor
wrap.anova(dv1 = bdata[c(6, 8)])

## ANOVA with 2 between-subjects factors
wrap.anova(dv1 = bdata$DV5, iv1 = bdata$IV1, iv2 = bdata$IV2)

## ANOVA with 1 within-subjects factor & 2 between-subjects factors
wrap.anova(dv1 = bdata[c(6, 8)], iv1 = bdata$IV1, iv2 = bdata$IV2)
}
\seealso{
\code{\link[ez]{ezANOVA}}
}
