% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.chi.R
\name{wrap.chi}
\alias{wrap.chi}
\title{Chi-square tests}
\usage{
wrap.chi(dv1, iv1 = NULL, p = rep(1/nlevels(dv1), nlevels(dv1)), correct = F)
}
\arguments{
\item{dv1}{Column vector containing the categorical dependent variable}

\item{iv1}{Column vector containing the categorical independent variable (contingency tests, only)}

\item{p}{A vector of probabilities representing expected cell frequencies (goodness-of-fit tests, only). By default, the function assumes equal expected cell frequencies for all levels of the dependent variable.}

\item{correct}{A logical argument: If FALSE, the function does not apply Yates's correction for 2x2 contingency tables; if TRUE, the function applies Yates's correction for 2x2 contingency tables.}
}
\description{
Performs one-way goodness-of-fit and two-way contingency tests. The
function delegates the primary operations to \code{\link[stats]{chisq.test}}.
}
\examples{
## One-way goodness-of-fit test
wrap.chi(dv1 = bdata$DV2)

## Two-way contingency test
wrap.chi(dv1 = bdata$DV2, iv1 = bdata$IV2)

}
\seealso{
\code{\link[stats]{chisq.test}}
}
