% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.desc.R
\name{wrap.desc}
\alias{wrap.desc}
\title{Descriptive statistics}
\usage{
wrap.desc(dv1, iv1 = NULL, iv2 = NULL)
}
\arguments{
\item{dv1}{Column vector containing the dependent variable}

\item{iv1, iv2}{Column vectors containing the independent variables}
}
\description{
Computes descriptive statistics for one dependent variable, parsed
by between 0 and 2 independent variables.
}
\examples{
## Parsing the dependent variable by 0 independent variables
wrap.desc(dv1 = bdata$DV5)

## Parsing the dependent variable by 1 independent variable
wrap.desc(dv1 = bdata$DV5, iv1 = bdata$IV2)

}
