% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.generate.R
\name{wrap.generate}
\alias{wrap.generate}
\title{Generate columns}
\usage{
wrap.generate(df, string1, string2, operation, newString = operation)
}
\arguments{
\item{df}{The data frame}

\item{string1, string2}{Character strings embedded in the names of corresponding columns}

\item{operation}{Character string to specify the operation (\code{"sum"}, \code{"difference"}, or
\code{"mean"})}

\item{newString}{Character string to replace \code{string1} and \code{string2} in the
outputted column/s. By default, \code{newString} equals the operation (\code{"sum"},
\code{"difference"}, or \code{"mean"}).}
}
\description{
Searches the data frame for pairs of columns whose names are identical
except for specific strings that you've embedded within them--such as "T1" versus "T2"
or "Participant1" versus "Participant2--and then generates new columns by computing
sums, differences, or means between pairs of corresponding columns. After generating
columns, the function assigns the revised data frame directly to the Global Environment.
}
\examples{
## Computing differences between T1 responses and T2 responses
wrap.generate(df = bdata, string1 = "T1", string2 = "T2", operation = "difference")

}
