% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.hist.R
\name{wrap.hist}
\alias{wrap.hist}
\title{Histogram plots}
\usage{
wrap.hist(
  dv1,
  likert = F,
  percent = T,
  binwidth = 1,
  ylim = NULL,
  ymajor = NULL,
  ylab = NULL,
  xlim = NULL,
  xmajor = NULL,
  xlab = NULL,
  title = NULL,
  size.axis.text.y = 12,
  size.axis.text.x = 12,
  size.title = 24
)
}
\arguments{
\item{dv1}{Column vector containing the dependent variable}

\item{likert}{A logical argument: If \code{FALSE}, the function assumes a
continuous dependent variable; if \code{TRUE}, the function assumes a discrete
dependent variable that takes on only integer values}

\item{percent}{A logical argument: if \code{TRUE}, the plot displays percentages
along the y axis; if \code{FALSE}, the plot displays counts}

\item{binwidth}{Numeric argument representing bin width}

\item{ylim}{Numeric vector containing y-axis limits}

\item{ymajor}{Numeric argument representing spacing of y-axis tick marks}

\item{ylab}{Character string containing the y-axis label}

\item{xlim}{Numeric vector containing x-axis limits}

\item{xmajor}{Numeric argument representing spacing of x-axis labels}

\item{xlab}{Character string containing the x-axis label}

\item{title}{Character string containing plot title}

\item{size.axis.text.x, size.axis.text.y, size.title}{Numeric arguments
containing font sizes}
}
\description{
Creates histogram plot for numerical dependent variables and prints
descriptive statistics in a summary table. The function delegates the primary
computations to \code{\link[ggplot2]{ggplot}}.
}
\examples{
## Histogram for a continuous dependent variable
wrap.hist(dv1 = bdata$DV1, likert = FALSE)

## Histogram for a discrete dependent variable
wrap.hist(dv1 = bdata$DV5, likert = TRUE)

}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
