% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.line.R
\name{wrap.line}
\alias{wrap.line}
\title{Line plots}
\usage{
wrap.line(
  dv1,
  iv1 = NULL,
  iv2 = NULL,
  reposition = NULL,
  rename1 = NULL,
  rename2 = NULL,
  rename3 = NULL,
  reorder1 = NULL,
  reorder2 = NULL,
  reorder3 = NULL,
  ylim = NULL,
  ymajor = NULL,
  ylab = NULL,
  xlab = NULL,
  title = NULL,
  size.axis.text.y = 12,
  size.axis.text.x = 16,
  size.title = 24,
  size.panel.title = 12,
  size.legend.text = 14
)
}
\arguments{
\item{dv1}{Multiple column vectors containing the within-subjects dependent variables}

\item{iv1, iv2}{Column vectors containing the independent variables}

\item{reposition}{Numeric vector to rearrange columns in the summary table and
thus reposition factors within the plot itself. For example, \code{reposition = c(1, 3, 2)}
reverses the order of the second and third columns in the summary table and
repositions the corresponding factors within the plot. The length of the reposition
vector must equal the total number of within-subjects and between-subjects factors.}

\item{rename1, rename2, rename3}{String vectors to rename the factor levels in the summary
table and thus rename factor levels within the plot itself. For example,
\code{rename1 = c("Close Friend", "Distant Stranger")} renames the levels in the first
column of the summary table and renames the corresponding factor levels within the plot.
(Note that the function applies the \code{reposition} argument to the summary table before applying
the \code{rename} arguments.)}

\item{reorder1, reorder2, reorder3}{String vectors to reorder the factor levels in the
summary table and thus reorder factor levels within the plot itself. For example,
\code{reorder1 = c("Stranger", "Friend")} reorders the levels in the first column of the
summary table and reorders the corresponding factor levels within the plot. (Note that
the function applies the \code{reposition} and \code{rename} arguments to the summary table
before applying the \code{reorder} arguments.)}

\item{ylim}{Numeric vector containing lower and upper y-axis limits}

\item{ymajor}{Numeric argument representing spacing of y-axis tick marks}

\item{ylab}{Character string containing the y-axis label}

\item{xlab}{Character string containing the x-axis label}

\item{title}{Character string containing the plot title}

\item{size.axis.text.y, size.axis.text.x, size.title, size.panel.title, size.legend.text}{Numeric
arguments containing font sizes}
}
\description{
Creates line plots for numerical dependent variables, adds error bars,
and prints descriptive statistics in a summary table. The function creates plots for
up to 3 factors total, including 1 within-subjects factor (corresponding to the path
of each line) and 0-2 between-subjects factors. The function delegates the primary
computations to \code{\link[ggplot2]{ggplot}}. Error bars are ±1 SE.
}
\examples{
## Line plot with 1 within-subjects factor
wrap.line(dv1 = bdata[c(6, 8)])

## Line plot with 1 within-subjects factor & 2 between-subjects factors
wrap.line(dv1 = bdata[c(6, 8)], iv1 = bdata$IV1, iv2 = bdata$IV2)

}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
