% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.lm.R
\name{wrap.lm}
\alias{wrap.lm}
\title{Linear regression}
\usage{
wrap.lm(model, standardized = FALSE)
}
\arguments{
\item{model}{The linear model}

\item{standardized}{A logical argument: if \code{FALSE}, the function
returns unstandardized coefficients; if \code{TRUE}, the function returns
standardized coefficients}
}
\description{
Performs linear regression analyses. The function delegates
the primary computations to \code{\link[stats]{lm}} and
\code{\link[sjstats]{std_beta}}. In the output, R^2 is not adjusted.
}
\examples{
## Linear regression with unstandardized coefficients
wrap.lm(model = bdata$DV7 ~ bdata$DV5 * bdata$DV6, standardized = FALSE)

# Linear regression with standardized coefficients
wrap.lm(model = bdata$DV7 ~ bdata$DV5 * bdata$DV6, standardized = TRUE)
}
\seealso{
\code{\link[stats]{lm}}, \code{\link[sjstats]{std_beta}}
}
