% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.rd.R
\name{wrap.rd}
\alias{wrap.rd}
\title{Round a number (without leading zeros)}
\usage{
wrap.rd(num, places = 2)
}
\arguments{
\item{num}{The number that will be rounded}

\item{places}{The number of decimal places to which to round the number}
}
\description{
Rounds a number, omits the leading zero, and retains at
least one significant digit when possible. (This is a helper function
that will be called by other functions in this package.)
}
\examples{
wrap.rd(num = 0.0145, places = 3)

}
\keyword{internal}
