% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.rd0.R
\name{wrap.rd0}
\alias{wrap.rd0}
\title{Round a number (with leading zeros)}
\usage{
wrap.rd0(num, places = 2)
}
\arguments{
\item{num}{The number that will be rounded}

\item{places}{The number of decimal places to which to round the number}
}
\description{
Rounds a number, includes the leading zero, and retains
at least one significant digit when possible. (This is a helper function
that will be called by other functions in this package.)
}
\examples{
wrap.rd0(num = 0.0145, places = 3)

}
\keyword{internal}
