% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.simple.R
\name{wrap.simple}
\alias{wrap.simple}
\title{ANOVA (simple main effects)}
\usage{
wrap.simple(dv1, iv1, iv2, adjustment = "none")
}
\arguments{
\item{dv1}{Column vector containing the dependent variable}

\item{iv1, iv2}{Column vectors containing the between-subjects independent variables.
The function will test for simple main effects of \code{iv1} separately at each level
of \code{iv2}.}

\item{adjustment}{Character string representing the method of p value adjustment
(\code{"none"}, \code{"holm"}, \code{"hochberg"}, \code{"hommel"},
\code{"bonferroni"}, \code{"BH"}, \code{"BY"}, or \code{"fdr"}).}
}
\description{
Computes simple main effects for a two-way, between-subjects ANOVA.
The function delegates the primary computations to \code{\link[phia]{testInteractions}}.
Note that this function assumes categorical (i.e., unordered) independent variables
and fixed effects. In the output, hp2 denotes partial eta squared.
}
\examples{
wrap.simple(dv1 = bdata$DV5, iv1 = bdata$IV1, iv2 = bdata$IV2)

}
\seealso{
\code{\link[phia]{testInteractions}}
}
