% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.t.ind.R
\name{wrap.t.ind}
\alias{wrap.t.ind}
\title{T test (independent samples)}
\usage{
wrap.t.ind(dv1, iv1, var.equal = T)
}
\arguments{
\item{dv1}{Column vector containing the dependent variable}

\item{iv1}{Column vector containing the independent variable}

\item{var.equal}{A logical argument: If TRUE, the function assumes equality of variances across conditions; if FALSE, the function does not assume equality of variances}
}
\description{
Performs independent-samples t tests. The function delegates
the primary computations to \code{\link[stats]{t.test}}. If variance differs
significantly by condition in your data, the function also displays the results
of Levene's test for equality of variances with centering at the median.
}
\examples{
wrap.t.ind(dv1 = bdata$DV5, iv1 = bdata$IV1)

}
\seealso{
\code{\link[stats]{t.test}}
}
