% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.t.one.R
\name{wrap.t.one}
\alias{wrap.t.one}
\title{T Test (One Sample)}
\usage{
wrap.t.one(dv1, mu = 0)
}
\arguments{
\item{dv1}{Column vector containing the dependent variable}

\item{mu}{Mean against which to compare responses on the dependent variable}
}
\description{
Performs one-sample t tests. The function delegates the primary
operations to \code{\link[stats]{t.test}}.
}
\examples{
wrap.t.one(dv1 = bdata$DV5, mu = 5)
}
\seealso{
\code{\link[stats]{t.test}}
}
