% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.t.pair.R
\name{wrap.t.pair}
\alias{wrap.t.pair}
\title{T test (paired samples)}
\usage{
wrap.t.pair(dv1, dv2, within_variation = TRUE)
}
\arguments{
\item{dv1, dv2}{Column vectors containing the dependent variables}

\item{within_variation}{A logical argument: If TRUE, the function uses within-participants variation when computing the effect size; if FALSE, the function does not use within-participants variation when computing the effect size.}
}
\description{
Performs paired-samples t tests. The function delegates the
primary computations to \code{\link[stats]{t.test}}.
}
\examples{
wrap.t.pair(dv1 = bdata$DV3_T1, dv2 = bdata$DV3_T2)

}
\seealso{
\code{\link[stats]{t.test}}
}
