% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.writeClipboard.R
\name{wrap.writeClipboard}
\alias{wrap.writeClipboard}
\title{Write to clipboard}
\usage{
wrap.writeClipboard(string)
}
\arguments{
\item{string}{The string to write to the clipboard}
}
\description{
Writes a string to the clipboard and removes the trailing newline. The
function delegates the primary operations to writeClipboard (package: utils). (This
is a helper function that will be called by other functions in this package.)
}
\examples{
wrap.writeClipboard(string = "text")

}
\seealso{
writeClipboard (package: utils)
}
