% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwrappers-package.R
\docType{package}
\name{bwrappers-package}
\alias{bwrappers}
\alias{bwrappers-package}
\title{bwrappers: Clean, analyze, and visualize data in the behavioral sciences}
\description{
This package contains wrappers and other functions for more
    efficiently cleaning, analyzing, and visualizing data in the behavioral
    sciences. The functions tailor syntax to the structure of a typical data
    frame in the behavioral sciences, automate intermediate steps such as
    splitting and reshaping data frames, return statistics in APA
    format, and copy these outputs directly to the clipboard.
}
\references{
\url{https://github.com/michaelkardas/bwrappers}
}
\author{
\strong{Maintainer}: Michael Kardas \email{mkardas@chicagobooth.edu}

}
\keyword{internal}
