% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.levene.R
\name{wrap.levene}
\alias{wrap.levene}
\title{Levene's test for equality of variances}
\usage{
wrap.levene(dv1, iv1, location = "median")
}
\arguments{
\item{dv1}{Column vector containing the dependent variable}

\item{iv1}{Column vector containing the independent variable}

\item{location}{Character string specifying the location from which
to compute the variance within each group (\code{"median"} or \code{"mean"})}
}
\description{
Performs Levene's test for equality of variances with one
between-subjects factor. The function delegates the primary computations
to \code{\link[ez]{ezANOVA}}. Note that this function does not use
bootstrapping and does not apply corrections. In the output, the F
statistic is equivalent to Levene's W, and hp2 denotes partial eta squared.
}
\examples{
wrap.levene(dv1 = bdata$DV5, iv1 = bdata$IV2, location = "median")

}
\seealso{
\code{\link[ez]{ezANOVA}}
}
