% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.t.ind.R
\name{wrap.t.ind}
\alias{wrap.t.ind}
\title{Independent-samples t test}
\usage{
wrap.t.ind(dv1, iv1, var.equal = T)
}
\arguments{
\item{dv1}{Column vector containing the dependent variable}

\item{iv1}{Column vector containing the independent variable}

\item{var.equal}{A logical argument: If TRUE, the function assumes equal variance across conditions; if FALSE, the function does not assume equal variance}
}
\description{
Performs independent-samples t tests. The function delegates
the primary computations to \code{\link[stats]{t.test}}.
}
\examples{
wrap.t.ind(dv1 = bdata$DV5, iv1 = bdata$IV1)

}
\seealso{
\code{\link[stats]{t.test}}
}
