% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.t.pair.R
\name{wrap.t.pair}
\alias{wrap.t.pair}
\title{Paired-samples t test}
\usage{
wrap.t.pair(dv1, dv2)
}
\arguments{
\item{dv1, dv2}{Column vectors containing the dependent variables}
}
\description{
Performs paired-samples t tests. The function delegates the
primary computations to \code{\link[stats]{t.test}}.
}
\examples{
wrap.t.pair(dv1 = bdata$DV3_T1, dv2 = bdata$DV3_T2)

}
\seealso{
\code{\link[stats]{t.test}}
}
